# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "dockm API is an HTTP API served by dockm. It is used by the\
    \ dockm UI and everything you can do with the UI can be done using the HTTP\
    \ API.\nYou can find out more about dockm at [http://dockm.io](http://dockm.io)\
    \ and get some support on [Slack](http://dockm.io/slack/).\n\n# Authentication\n\
    \nMost of the API endpoints require to be authenticated as well as some level\
    \ of authorization to be used.\ndockm API uses JSON Web Token to manage authentication\
    \ and thus requires you to provide a token in the **Authorization** header of\
    \ each request\nwith the **Bearer** authentication mechanism.\n\nExample:\n```\n\
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE\n\
    ```\n\n# Security\n\nEach API endpoint has an associated access policy, it is\
    \ documented in the description of each endpoint.\n\nDifferent access policies\
    \ are available:\n* Public access\n* Authenticated access\n* Restricted access\n\
    * Administrator access\n\n### Public access\n\nNo authentication is required to\
    \ access the endpoints with this access policy.\n\n### Authenticated access\n\n\
    Authentication is required to access the endpoints with this access policy.\n\n\
    ### Restricted access\n\nAuthentication is required to access the endpoints with\
...
    \ with this access policy.\n"
...
  contact:
...
host: "dockm.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /dockerhub:
    get:
      tags:
      - "dockerhub"
      summary: "Retrieve DockerHub information"
      description: "Use this endpoint to retrieve the information used to connect\
        \ to the DockerHub  \n**Access policy**: authenticated\n"
      operationId: "DockerHubInspect"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/DockerHubInspectResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "dockerhub"
      summary: "Update DockerHub information"
      description: "Use this endpoint to update the information used to connect to\
        \ the DockerHub  \n**Access policy**: administrator\n"
      operationId: "DockerHubUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "DockerHub information"
        required: true
        schema:
          $ref: "#/definitions/DockerHubUpdateRequest"
      responses:
        200:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  DockerHubInspectResponse:
    type: "object"
    properties:
      Authentication:
        type: "boolean"
        example: true
        description: "Is authentication against DockerHub enabled"
      Username:
        type: "string"
        example: "hub_user"
        description: "Username used to authenticate against the DockerHub"
      Password:
        type: "string"
        example: "hub_password"
        description: "Password used to authenticate against the DockerHub"
  DockerHubUpdateRequest:
    type: "object"
    required:
    - "Authentication"
    - "Password"
    - "Username"
    properties:
      Authentication:
        type: "boolean"
        example: true
        description: "Enable authentication against DockerHub"
      Username:
        type: "string"
        example: "hub_user"
        description: "Username used to authenticate against the DockerHub"
      Password:
        type: "string"
        example: "hub_password"
        description: "Password used to authenticate against the DockerHub"
...